requirements.txt
streamlit
audio-recorder-streamlit
openai
python-dotenv
openai-whisper
pyperclip